backend/pyproject.toml
[project]
name = "ape-replica"
version = "0.1.0"
description = "Autonomous research infrastructure for AI governance"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "aiosqlite>=0.20.0",
    "alembic>=1.14.0",
    "pydantic-settings>=2.7.0",
    "trueskill>=0.4.5",
    "anthropic>=0.42.0",
    "openai>=1.60.0",
    "google-genai>=1.0.0",
    "httpx>=0.28.0",
    "pyyaml>=6.0.2",
    "apscheduler>=3.10.4",
    "python-multipart>=0.0.18",
    "greenlet>=3.0.0",
    "asyncpg>=0.30.0",
    "slowapi>=0.1.9",
    "tenacity>=9.0.0",
    "sentry-sdk[fastapi]>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
